// generated/src/app/api-models/smformalparameter-s.model.ts
/* Copyright (c) 2020 . All Rights Reserved. */

export class SmformalparameterS {
  namedobjectIdSequenceid: number;
  namedobjectIdObjecttype: number;
  formalparameterParametername: string;
  formalparameterDescription: string;
  formalparameterIsscalar: number;
  formalparameterIsmandatory: number;
  formalparameterUserdefflag: number;
  numberformalparameterLbound: number;
  numberformalparameterUbound: number;
  objectformalparameterFpot: number;
}